Added gitconfig and fixed install.sh creating recursive bins.

+++ install.sh
@@ -3,5 +3,8 @@
 ln -s ~/.dotfiles/bashrc ~/.bashrc
 ln -s ~/.dotfiles/bash_profile ~/.bash_profile
 ln -s ~/.dotfiles/vimrc ~/.vimrc
+ln -s ~/.dotfiles/gitconfig ~/.gitconfig
 
-ln -s ~/.dotfiles/bin ~/bin
+if [[ ! -e ~/bin ]]; then
+	ln -s ~/.dotfiles/bin ~/bin
+fi
